# api-spec.yaml
openapi: 3.0.3
info:
  title: Pet Clinic API
  description: API para gerenciamento de Users, Clients, Owners, Pets e Appointments com máquina de estados.
  version: 1.0.0
  contact:
    name: Construção de Software PUCRS
    email: contato@exemplo.com
servers:
  - url: http://localhost:3000/api/v1
    description: Servidor de desenvolvimento local
  - url: https://seu-app.herokuapp.com/api/v1
    description: Servidor de produção
security:
  - bearerAuth: []
...
paths:
...
  /owners:
    get:
      tags: [Owners]
      summary: Listar owners
      operationId: listOwners
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de owners
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponseOwners'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Owners]
      summary: Criar owner
      operationId: createOwner
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOwnerRequest'
      responses:
        '201':
          description: Owner criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /owners/{ownerId}:
    parameters:
      - in: path
        name: ownerId
        required: true
        schema:
          type: string
    get:
      tags: [Owners]
      summary: Buscar owner por ID
      operationId: getOwnerById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Owner encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Owners]
      summary: Atualizar owner
      operationId: updateOwner
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOwnerRequest'
      responses:
        '200':
          description: Owner atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Owner'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Owners]
      summary: Deletar owner
      operationId: deleteOwner
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Owner deletado com sucesso
        '409':
          description: Não é permitido deletar owner com pets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /pets/{petId}:
    parameters:
      - in: path
        name: petId
        required: true
        schema:
          type: string
    get:
      tags: [Pets]
      summary: Buscar pet por ID
      operationId: getPetById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pet encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Pets]
      summary: Atualizar pet
      operationId: updatePet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePetRequest'
      responses:
        '200':
          description: Pet atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Pets]
      summary: Deletar pet
      operationId: deletePet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pet deletado com sucesso
        '409':
          description: Não é permitido deletar pet com appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /appointments:
    get:
      tags: [Appointments]
      summary: Listar appointments
      description: Filtrar por petId, ownerId, date; paginação via limit/offset
      operationId: listAppointments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: petId
          schema:
            type: string
        - in: query
          name: ownerId
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
            format: date
            example: "2025-12-10"
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponseAppointments'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Appointments]
      summary: Criar appointment (agendar)
      operationId: createAppointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
      responses:
        '201':
          description: Appointment criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Pet or Owner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          $ref: '#/components/responses/Conflict'
  /appointments/{appointmentId}:
    parameters:
      - in: path
        name: appointmentId
        required: true
        schema:
          type: string
    get:
      tags: [Appointments]
      summary: Buscar appointment por ID
      operationId: getAppointmentById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Appointment encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Appointments]
      summary: Atualizar appointment
      operationId: updateAppointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointmentRequest'
      responses:
        '200':
          description: Appointment atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Appointments]
      summary: Deletar appointment
      operationId: deleteAppointment
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Appointment deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Acesso não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unauthorized:
              value:
                error: "Unauthorized"
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid:
              value:
                error: "Invalid input data"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            not_found:
              value:
                error: "Not found"
    Conflict:
      description: Conflito (p.ex. recurso duplicado ou regra de negócio)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            conflict:
              value:
                error: "Conflict"
    InternalError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            server_error:
              value:
                error: "internal"
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid input data"
        details:
          type: string
          example: "Campo X é obrigatório"
    PaginatedResponseBase:
      type: object
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          example: 100
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0
        hasMore:
          type: boolean
          example: true

    # ------------------------
    # Users (admin)
    # ------------------------
...
    Owner:
      type: object
      required:
        - ownerId
        - name
        - email
        - phone
        - createdAt
      properties:
        ownerId:
          type: string
          example: "o1a2b3c4-1234-5678-9abc-def012345678"
        name:
          type: string
          example: "Ana Laura Lopes"
        email:
          type: string
          format: email
          example: "ana@example.com"
        phone:
          type: string
          example: "+55 51 99999-9999"
        address:
          type: string
          example: "Rua Exemplo, 123"
        createdAt:
          type: string
          format: date-time
    CreateOwnerRequest:
      type: object
      required:
        - name
        - email
        - phone
      properties:
        name: { type: string }
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
    UpdateOwnerRequest:
      type: object
      properties:
        name: { type: string }
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
    PaginatedResponseOwners:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseBase'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Owner'

    # ------------------------
    # Pets
    # ------------------------
    Pet:
      type: object
      required:
        - petId
        - ownerId
        - name
        - species
        - createdAt
      properties:
        petId:
          type: string
          example: "p1a2b3c4-9876-5432-1abc-def012345678"
        ownerId:
          type: string
          example: "o1a2b3c4-1234-5678-9abc-def012345678"
        name:
          type: string
          example: "Rex"
        species:
          type: string
          example: "cachorro"
        breed:
          type: string
          example: "Labrador"
        birthDate:
          type: string
          format: date
          example: "2018-05-20"
        weight:
          type: number
          example: 12.5
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
...
    UpdatePetRequest:
      type: object
      properties:
        ownerId:
          type: string
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        birthDate:
          type: string
          format: date
        weight:
          type: number
        notes:
          type: string
...
    AppointmentStatus:
      type: string
      description: Estado atual do appointment
      enum:
        - pendente
        - confirmado
        - cancelado
        - concluido
      example: "pendente"
...
    Appointment:
      type: object
      required:
        - appointmentId
        - petId
        - ownerId
        - date
        - time
        - type
        - status
        - createdAt
      properties:
        appointmentId:
          type: string
          example: "a1b2c3d4-1234-5678-9abc-defghijklmn1"
        petId:
          type: string
          example: "p1a2b3c4-9876-5432-1abc-def012345678"
        ownerId:
          type: string
          example: "o1a2b3c4-1234-5678-9abc-def012345678"
        date:
          type: string
          format: date
          example: "2025-12-10"
        time:
          type: string
          pattern: "^([01]\\d|2[0-3]):([0-5]\\d)$"
          example: "14:30"
        type:
          type: string
          example: "consulta"
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        veterinarian:
          type: string
          example: "Dr. Fulano"
        notes:
          type: string
          example: "Trazer vacinas anteriores"
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        canceledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
    CreateAppointmentRequest:
      type: object
      required:
        - petId
        - ownerId
        - date
        - time
        - type
      properties:
        petId:
          type: string
        ownerId:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
          pattern: "^([01]\\d|2[0-3]):([0-5]\\d)$"
        type:
          type: string
          description: Tipo de atendimento (consulta, vacina, banho, tosa, etc)
        status:
          $ref: '#/components/schemas/AppointmentStatus'
          description: Estado inicial (recomendado 'pendente')
        veterinarian:
          type: string
        notes:
          type: string
    UpdateAppointmentRequest:
      type: object
      properties:
        petId:
          type: string
        ownerId:
          type: string
        date:
          type: string
          format: date
        time:
          type: string
          pattern: "^([01]\\d|2[0-3]):([0-5]\\d)$"
        type:
          type: string
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        veterinarian:
          type: string
        notes:
          type: string
...
    PaginatedResponseAppointments:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseBase'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Appointment'
...